refactor: component renaming
