--- src/features/Product/index.ts
@@ -1,7 +1,0 @@
-export { getSelectedError } from './model/selectors/productSelectors';
-
-export { productReducer } from './model/slices/productSlice';
-
-export { SelectedProduct } from './ui/ProductSelected';
-
-export { ProductSchema } from './model/types/productSelectedTypes';
