+++ src/features/SelectedProduct/index.ts
@@ -1,0 +1,7 @@
+export { SelectedProduct } from './ui/SelectedProduct';
+
+export { getSelectedError } from './model/selectors/selectedProductSelectors';
+
+export { productReducer } from './model/slices/selectedProductSlice';
+
+export { ProductSchema } from './model/types/selectedProductTypes';
