+++ src/features/SelectedProduct/model/selectors/selectedProductSelectors.ts
