+++ src/features/SelectedProduct/model/services/getSelectedProductByKey.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { ErrorWithResponse, Images, Product } from 'shared/types';
-import { ProductReject, ProductSelectedResponse } from '../types/productSelectedTypes';
+import { ProductReject, ProductSelectedResponse } from '../types/selectedProductTypes';
 import { setPrices } from 'shared/lib/changeData';
 
 const PROJECT_KEY = process.env.PROJECT_KEY;
