+++ src/features/SelectedProduct/model/types/selectedProductTypes.ts
