+++ src/pages/product/Product.tsx
@@ -1,7 +1,7 @@
-import { SelectedProduct, getSelectedError } from 'features/Product';
+import { SelectedProduct, getSelectedError } from 'features/SelectedProduct';
 import { NotFound } from 'pages/notFound/NotFound';
 import { FC } from 'react';
-import { useAppSelector } from 'shared/lib/hooks/useAppSelect/useAppSelect';
+import { useAppSelector } from 'shared/lib/hooks';
 
 export const Product: FC = () => {
   const error = useAppSelector(getSelectedError);
